Moved listener to mainviewmodel
